Update callers of createAzureCommunicationCallAdapter
