+++ packages/storybook/stories/QuickStarts/snippets/QuickstartCompositeComplete.snippet.tsx
@@ -36,7 +36,12 @@ function App(): JSX.Element {
         )
       );
       setCallAdapter(
-        await createAzureCommunicationCallAdapter({ communicationUserId: userId }, token, { groupId }, displayName)
+        await createAzureCommunicationCallAdapter(
+          { kind: 'communicationUser', communicationUserId: userId },
+          displayName,
+          new AzureCommunicationTokenCredential(token),
+          { groupId }
+        )
       );
     };
     createAdapter();
