--- samples/Calling/src/app/utils/refreshToken.ts
@@ -1,19 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { AbortSignalLike } from '@azure/core-http';
-
-const postRefreshTokenParameters = {
-  method: 'POST'
-};
-
-export const refreshTokenAsync = (userIdentity: string): ((abortSignal?: AbortSignalLike) => Promise<string>) => {
-  return async (): Promise<string> => {
-    const response = await fetch(`/refreshToken/${userIdentity}`, postRefreshTokenParameters);
-    if (response.ok) {
-      return (await response.json()).token;
-    } else {
-      throw new Error('could not refresh token');
-    }
-  };
-};
